+++ src/include/CodeGenerator.hpp
@@ -4,7 +4,7 @@
 #include <cstdio>
 #include <string>
 
-#include "list.hpp"
+#include "List.hpp"
 
 namespace njnr
 {
